Refactored executor tests
